Fix conversion warning in tst_qresultstore
Changed type of `id`, `int` -> `size_t`, to match `liveCount`, in
`CountedObject` struct.

Fixes: QTBUG-122301

+++ tests/auto/corelib/thread/qresultstore/tst_qresultstore.cpp
@@ -549,7 +549,7 @@ struct CountedObject
         const size_t previousLiveCount;
     };
 
-    int id = 0;
+    size_t id = 0;
     static size_t liveCount;
 };
